[Mobile]Update video player to open the URL by browser on Android (#16089)
* Update video player to open the URL by browser

* Render play button even if video is not ready to play yet

* Add/update alert messages

* Revert back warn to error

* Do not show play button until video player has a height

* Resolve merge conflicts that re-introduced a removed file
