+++ packages/block-library/src/video/video-player.native.scss
@@ -13,7 +13,7 @@ $play-icon-size: 50;
 	align-items: center;
 	align-self: center;
 	position: absolute;
-	background-color: #e9eff3;
+	background-color: transparent;
 	opacity: 0.3;
 }
